Use BigDecimal for monetary values

+++ hotel-pullman-service/src/main/java/com/hotel/model/Chambre.java
@@ -1,17 +1,19 @@
 package com.hotel.model;
 
+import java.math.BigDecimal;
+
 import java.time.LocalDate;
 
 public class Chambre {
     private String numero;
     private int nombreLits;
-    private double prixParNuit;
+    private BigDecimal prixParNuit;
     private LocalDate dateDebutDisponible;
     private LocalDate dateFinDisponible;
     private String imagePath;
 
 
-    public Chambre(String numero, int nombreLits, double prixParNuit, LocalDate dateDebutDisponible, LocalDate dateFinDisponible, String imagePath) {
+    public Chambre(String numero, int nombreLits, BigDecimal prixParNuit, LocalDate dateDebutDisponible, LocalDate dateFinDisponible, String imagePath) {
         this.numero = numero;
         this.nombreLits = nombreLits;
         this.prixParNuit = prixParNuit;
@@ -28,7 +30,7 @@ public class Chambre {
         return nombreLits;
     }
     
-    public double getPrixParNuit() {
+    public BigDecimal getPrixParNuit() {
         return prixParNuit;
     }
